requirements.txt
numpy==1.22.3
pandas==1.3.5
scikit-learn==1.0.2
xgboost==1.6.0
hydra-core==1.1.2
hydra_colorlog==6.6.0
kaggle==1.5.12
seaborn==0.11.2
mlflow==1.25.1
